requirements.txt
joblib==1.5.1
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
plotly==6.5.0
PyYAML==6.0.3
scikit_learn==1.7.2
seaborn==0.13.2
streamlit==1.52.1
streamlit_option_menu==0.4.0
# tensorflow==2.19.0   # used locally for model training, not needed for deployment
